refactor(providers): extract AlibabaCodingPlan cookie importer behind shared OneConsole type
The Qwen Cloud provider already piggybacks on AlibabaCodingPlanCookieImporter's generic importSession(browserDetection:domains:isAuthenticatedSession:logPrefix:sessionLabel:logger:) entry point. Today that generic overload lives inside the Alibaba module path, which makes the shared-ness implicit and forces Qwen to reach into a sibling provider's namespace.

Move the generic implementation -- SessionInfo, importSession, hasSession, cookieImportCandidates, matchesCookieDomain, normalizeCookieDomain -- into a new Providers/Shared/AliyunOneConsole/AliyunOneConsoleCookieImporter.swift. Keep AlibabaCodingPlanCookieImporter as a thin Alibaba-specific wrapper that declares the cookie domains and ticket predicate, and re-export SessionInfo as a typealias so existing call sites compile unchanged.

QwenCloudCookieImporter now calls AliyunOneConsoleCookieImporter directly instead of going through the Alibaba namespace.

The Alibaba-specific Chromium fallback importer (AlibabaChromiumCookieFallbackImporter) stays in the Alibaba module because it carries Alibaba ticket-name knowledge. AliyunOneConsoleCookieImporter delegates to it.

No behavior change.

Verified:
- swift test --filter AlibabaCodingPlan: 58 tests passed
- swift test --filter AlibabaTokenPlan: 37 tests passed
- swift test --filter QwenCloud: 27 tests passed
- make check: 0 violations
